Bugfix for nexted rules returning sentinels
Using an `R` rule with a nested `Rule` that returned a `Sentinel` resulted in an unexpected exception. This was due to the lack of handling of sentinels by `R.query()`. This has been addressed by, checking the result of the nested `R.query` and:
* returning the `EMPTY` sentinel without further processing of the `R` rule if a nested rule returns `EMPTY`,
* replacing the `UNIVERSAL` sentinel with an empty `Q` and returning the `UNIVERSAL` sentinel if the accumulated `Q` is empty.

This ensures `R` is compliant with the both the expected return value of nested rules, and returns the expected result (using sentinels where appropriate) as used by the DRF `RulePermission` class and `Rule.filter()` method.