Fix R reverse relationship traversal crash (#3)
R would fail when attempting to follow a reverse relationship (i.e. the
'other' side of a `ForeignKey`/`OneToOneField`) and there are no
record(s) at the other side of the relationship for the model instance
being checked. This occured as accessing the reverse relationship
attribute would raise an exception instead of returning `None`.

`None` is forced as the LHS of the equation, relying on the existing
`Rule` subclass `None`-handling behaviour.

Also fixed test environment. Necessitated ending support for Python 3.5.