Use `filter()` for `Rule.check()` instead of trying to emulate the Django ORM (#4)
This is a large deviation from how upstream Bridgekeeper works, and is justified on the premise that `check()` can never accurately/reliably emulate the Django ORM, which is what Bridgekeeper attempts to achieve by promising `Rule`-consistent behaviour between `check()` (single instance) and `filter()` (filtering a `QuerySet`). For example, `R.check()` does not handle traversing relationships in a way that is consistent with `filter()`.

With this PR, we move to having `check()` call `filter()` under the hood. It pre-filters the `QuerySet` so that it only contains the provided instance, and returns whether or not the `QuerySet` contains any records (i.e. `QuerySet.exists()`).

This is an initial implementation that passes al of the current tests. At a minimum, documentation still needs to be updated. I also believe that there are a lot of unit tests that are no longer relevant as they're all checking what is now a single implementation of `check()`. It may be worth keeping around them for now.

I also had to leave the current implementation of  `blanket_rule.check()` in order for a blanket rule test to pass. I think that this is expected given the nature of `blanket_rule()` (i.e. it doesn't take an instance context, only a user context, so a discriminating `filter()` doesn't make sense at all). As such, `blanket_rule.query()` simply returns `UNIVERSAL` or `EMPTY` depending on whether or not the rule function returns `True` or `False`. Since the (new) base implementation of `Rule.check()` always returns `False` if an instance is not provided, the base implementation needs to be sidestepped in situations where an instance is not provided...which I think only happens with blanket rules?